Create ccd_script.py
Preliminary script for creating color-color diagrams with Alcala+2017 X-Shooter data matched with Spitzer photometry. Something is wrong with this iteration of it: the flux ratios for H-alpha when converting to magnitudes from flux densities are negative and very very small (e.g. -5*10**(-48)), so there must be something wrong with the way I'm calculating the flux densities.